+++ src/main/java/org/betonquest/betonquest/menu/betonquest/MenuEvent.java
@@ -4,36 +4,27 @@ import org.betonquest.betonquest.api.common.function.QuestConsumer;
 import org.betonquest.betonquest.api.profile.OnlineProfile;
 import org.betonquest.betonquest.api.quest.QuestException;
 import org.betonquest.betonquest.api.quest.event.online.OnlineEvent;
-import org.betonquest.betonquest.instruction.variable.Variable;
-import org.betonquest.betonquest.menu.MenuID;
-import org.betonquest.betonquest.menu.RPGMenu;
-import org.jetbrains.annotations.Nullable;
 
 /**
- * Event to open or close menus.
+ * Event to open, close or update menus.
  */
 public class MenuEvent implements OnlineEvent {
     /**
-     * The stuff to do with the profile.
+     * The action to do with the profile.
      */
-    private final QuestConsumer<OnlineProfile> doStuff;
+    private final QuestConsumer<OnlineProfile> action;
 
     /**
      * Creates a new MenuQuestEvent.
      *
-     * @param rpgMenu the rpg menu instance to open and close menus
-     * @param menuID  the menu id to open or null if open menus should be closed
+     * @param action the action to do with the profile
      */
-    public MenuEvent(final RPGMenu rpgMenu, @Nullable final Variable<MenuID> menuID) {
-        if (menuID != null) {
-            doStuff = profile -> rpgMenu.openMenu(profile, menuID.getValue(profile));
-        } else {
-            doStuff = RPGMenu::closeMenu;
-        }
+    public MenuEvent(final QuestConsumer<OnlineProfile> action) {
+        this.action = action;
     }
 
     @Override
     public void execute(final OnlineProfile profile) throws QuestException {
-        doStuff.accept(profile);
+        action.accept(profile);
     }
 }